Time: 241 ms (9.45%), Space: 14.5 MB (23.94%) - LeetHub

+++ 264-ugly-number-ii/264-ugly-number-ii.cpp
@@ -22,6 +22,6 @@ public:
             c++;
             
         }
-        return (int)mini;
+        return mini;
     }
 };
